refactor: move usart init to its module
